refactor : Product 도메인 분리 및 JPA 적용
