+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,0 +1,30 @@
+package kitchenpos.product.application;
+
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductCreateRequest;
+import kitchenpos.product.dto.ProductResponse;
+import kitchenpos.product.repository.ProductRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+public class ProductService {
+    private final ProductRepository productRepository;
+
+    public ProductService(final ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Transactional
+    public ProductResponse create(final ProductCreateRequest request) {
+        Product product = new Product(request.getName(), request.getPrice());
+        Product savedProduct = productRepository.save(product);
+        return ProductResponse.of(savedProduct);
+    }
+
+    public List<ProductResponse> list() {
+        return ProductResponse.list(productRepository.findAll());
+    }
+}
