+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.product.repository;
+
+import kitchenpos.product.domain.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
